Extract `BitOrder` into `"bitstreams/BitStreams.h"`, fix includes
