+++ cmake/Modules/cpu-large-page-size.cpp
@@ -1,3 +1,4 @@
+#include <cstddef>
 #include <iostream>
 
 #if defined(__i386__) || defined(__x86_64__)
