+++ src/librawspeed/decoders/KdcDecoder.cpp
@@ -23,7 +23,7 @@
 #include "adt/Casts.h"
 #include "adt/NORangesSet.h"
 #include "adt/Point.h"
-#include "common/Common.h"
+#include "bitstreams/BitStreams.h"
 #include "common/RawImage.h"
 #include "decoders/RawDecoderException.h"
 #include "decompressors/UncompressedDecompressor.h"
