+++ src/librawspeed/decoders/MrwDecoder.cpp
@@ -22,7 +22,7 @@
 
 #include "decoders/MrwDecoder.h"
 #include "adt/Point.h"
-#include "common/Common.h"
+#include "bitstreams/BitStreams.h"
 #include "common/RawImage.h"
 #include "decoders/RawDecoderException.h"
 #include "decompressors/UncompressedDecompressor.h"
