+++ src/librawspeed/decoders/OrfDecoder.cpp
@@ -27,6 +27,7 @@
 #include "adt/NORangesSet.h"
 #include "adt/Point.h"
 #include "bitstreams/BitStreamerMSB.h"
+#include "bitstreams/BitStreams.h"
 #include "common/Common.h"
 #include "common/RawImage.h"
 #include "decoders/RawDecoderException.h"
