+++ src/librawspeed/decoders/PefDecoder.cpp
@@ -25,6 +25,7 @@
 #include "adt/Casts.h"
 #include "adt/Optional.h"
 #include "adt/Point.h"
+#include "bitstreams/BitStreams.h"
 #include "common/Common.h"
 #include "common/RawImage.h"
 #include "decoders/RawDecoderException.h"
