+++ src/librawspeed/decoders/ThreefrDecoder.cpp
@@ -22,7 +22,7 @@
 #include "decoders/ThreefrDecoder.h"
 #include "adt/Casts.h"
 #include "adt/Point.h"
-#include "common/Common.h"
+#include "bitstreams/BitStreams.h"
 #include "common/RawImage.h"
 #include "decoders/RawDecoderException.h"
 #include "decompressors/HasselbladLJpegDecoder.h"
